+++ dash/dash-renderer/src/types/callbacks.ts
@@ -1,3 +1,5 @@
+import {PatchAnalysis} from '../actions/patchAnalysis';
+
 type CallbackId = string | {[key: string]: any};
 
 export interface ICallbackDefinition {
@@ -85,6 +87,11 @@ export type CallbackResult = {
     data?: CallbackResponse;
     error?: Error;
     payload: ICallbackPayload | null;
+    patchedOutputs?: PatchedOutputs;
+};
+
+export type PatchedOutputs = {
+    [idStr: string]: PatchAnalysis;
 };
 
 export type BackgroundCallbackInfo = {